sideLinksChanges
