+++ eds/templates/stories/stories.js
@@ -1,20 +1,5 @@
 import { buildBlock, getMetadata } from '../../scripts/aem.js';
-import { div, p } from '../../scripts/dom-builder.js';
 import { buildArticleSchema, buildPodcastEpisodeSchema } from '../../scripts/schema.js';
-
-function sideLinksDiv(linkHeading) {
-  const main = document.querySelector('main');
-  const divEl = div({ class: 'sidelinks leading-5 text-sm font-bold text-black flex flex-col gap-y-3' });
-  const allParagraphs = main.querySelectorAll('p a[title="link"]');
-  if (allParagraphs.length > 0) {
-    divEl.append(p(linkHeading));
-    allParagraphs.forEach((elPara) => {
-      elPara.classList.add(...'border-b border-b-gray-300 py-2 mx-0 w-auto mt-2 pb-3'.split(' '));
-      divEl.append(p({ class: 'leading-5 text-base font-medium text-[#378189]' }, elPara));
-    });
-  }
-  return divEl;
-}
 
 export default async function buildAutoBlocks() {
   const main = document.querySelector('main');
@@ -29,7 +14,7 @@ export default async function buildAutoBlocks() {
   sectionMiddle.prepend(
     buildBlock('story-info', { elems: [] }),
     buildBlock('social-media', { elems: [] }),
-    sideLinksDiv('Explore our products'),
+    buildBlock('sidelinks', { elems: [] }),
   );
 }
